Dividing Slider into separate class. For future: try to swap aggregation between Waveform and Slider

+++ src/MainWindow.h
@@ -17,5 +17,5 @@ private:
 	boost::signals2::connection  SliderInteraction;
 
 	bool AudioFileHandler (std::string FileName);
-	void SliderHandler (int CenterSample);
+	void SliderHandler (double CenterSample);
 };
